Extract background and terrain generation code
Extract background and terrain block generation functions into a separate file and remove player and enemy related code.

* Add `background_and_terrain.py` with `get_background` and `get_block` functions.
* Remove `get_background` and `get_block` functions from `tutorial.py`.
* Import `get_background` and `get_block` functions from `background_and_terrain` in `tutorial.py`.
* Remove all player and enemy related code from `tutorial.py`.
